backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
scikit-learn>=1.3.0
ta>=0.10.0
httpx>=0.24.0
websockets>=11.0.0
python-jose>=3.3.0
passlib>=1.7.4
matplotlib>=3.7.0
langchain>=0.0.267
openai>=0.27.8 